FinalProject/requirements.txt
altgraph==0.17.2
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
bootstrap4==0.1.0
certifi==2021.10.8
chardet==3.0.4
click==8.0.3
colorama==0.4.4
distlib==0.3.3
Django==3.2.9
django-cors-headers==3.10.1
django-extensions==3.1.5
django-filter==21.1
django-widget-tweaks==1.4.9
djangorestframework==3.12.4
et-xmlfile==1.1.0
filelock==3.3.2
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.2
idna==2.6
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysqlclient==2.1.0
openpyxl==3.0.9
pefile==2021.9.3
Pillow==8.4.0
platformdirs==2.4.0
pycparser==2.21
pyinstaller==4.6
pyinstaller-hooks-contrib==2021.3
pyparsing==3.0.6
python-dateutil==2.7.3
python-stdnum==1.17
pytz==2019.1
pywin32==302
pywin32-ctypes==0.2.0
requests==2.21.0
requests-toolbelt==0.9.1
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.27
sqlparse==0.4.2
tk==0.1.0
urllib3==1.24.3
virtualenv==20.10.0
Werkzeug==2.0.2
zope.event==4.5.0
zope.interface==5.4.0
